Added IsSessionActive property to character_session
No session means character isn't in the world. A session existing means
it is/was in the world. Sessions can be reclaimed, meaing you can rejoin
transient instances properly. Sessions can be active or inactive, active
means it is claimed and no other use can log in.

There is a race condition here. Though it is a known and handlable one.
Meaning multiple ongoing session claims could be occuring. But only one
will ever be able to be claimed.