Added more AllModel Tests; Major Response Code Change
Added change to response codes which make it fail by default. I
identifable a potential critical design fault resulting in future
problems that could arrise with the default value 0 representing
success. It is better for the default ctor of the DTO to be fail. Even
if it's never explictly used. Errors or other faults could cause it to
happen and we don't want the problems to propagate since server to
server communication will need to trust the response code.