Added hack for interest leaving on player disconnect
This is NOT a final solution, but it is what will work for now. There
are afew design flaws, and this is one of them. Will address at a later
point along with the others. This one took a long time to even come up
with a hack for. So, I don't expect to solve this soon.