Added initial naive implementation for FieldValueUpdateFactory
We MUST come up with a copyless, allocationless, threadsafe, lockless
and serializable solution. Much of the Server CPU time will be spent
serializing and copying these bytes around otherwise in MASSIVE scale
scenarios.

It's absolutely CRITICAL that we solve this issue one day for
performance of the system at scale. This is one of the most important
packets and hot paths of the backend.