Fix other existing race condition in session cleanup
Sessions cleanup had another race condition where the player
disconnected during the claim handlers attempt to load character data
from the gameserver. This made it possible for the cleanup tickable on
the main thread to service the deconstruction, which would find no
entity.

This makes it so that entities after creation are rechecked for
connection state and a cleanup is queued up if the client is
disconnected.