Added PreBurstStartInstanceClientConnection
This component handles finding the zone/instance a client should connect
to, and then it connects to it and starts pumping messages. However,
this component and scene is liklely short lived. It will try to auth the
session and then the server is likely to tell it to load a new scene.